+++ packages/core/src/layers/PageLayer.tsx
@@ -71,6 +71,8 @@ export const PageLayer: React.FC<{
     });
     const [canvasLayerRendered, setCanvasLayerRendered] = React.useState(false);
     const [textLayerRendered, setTextLayerRendered] = React.useState(false);
+    const canvasLayerRef = React.useRef<HTMLCanvasElement>();
+    const textLayerRef = React.useRef<HTMLDivElement>();
 
     const { page, pageHeight, pageWidth } = pageSize;
 
@@ -193,6 +195,7 @@ export const PageLayer: React.FC<{
                             attrs: {},
                             children: (
                                 <CanvasLayer
+                                    canvasLayerRef={canvasLayerRef}
                                     height={h}
                                     page={page}
                                     pageIndex={pageIndex}
@@ -220,6 +223,7 @@ export const PageLayer: React.FC<{
                             attrs: {},
                             children: (
                                 <TextLayer
+                                    containerRef={textLayerRef}
                                     page={page}
                                     pageIndex={pageIndex}
                                     plugins={plugins}
@@ -238,12 +242,14 @@ export const PageLayer: React.FC<{
                         plugin.renderPageLayer ? (
                             <React.Fragment key={idx}>
                                 {plugin.renderPageLayer({
+                                    canvasLayerRef,
                                     canvasLayerRendered,
                                     doc,
                                     height: h,
                                     pageIndex,
                                     rotation: rotationValue,
                                     scale,
+                                    textLayerRef,
                                     textLayerRendered,
                                     width: w,
                                 })}