+++ packages/core/src/types/Plugin.ts
@@ -44,6 +44,7 @@ export interface PluginOnCanvasLayerRender {
 }
 
 export interface PluginRenderPageLayer {
+    canvasLayerRef: React.MutableRefObject<HTMLCanvasElement>;
     // Is the canvas layer rendered completely?
     canvasLayerRendered: boolean;
     doc: PdfJs.PdfDocument;
@@ -51,6 +52,7 @@ export interface PluginRenderPageLayer {
     pageIndex: number;
     rotation: number;
     scale: number;
+    textLayerRef: React.MutableRefObject<HTMLDivElement>;
     // Is the text layer rendered completely?
     textLayerRendered: boolean;
     width: number;