+++ packages/highlight/src/HighlightAreaList.tsx
@@ -9,8 +9,8 @@
 import type { Store } from '@react-pdf-viewer/core';
 import * as React from 'react';
 import { HighlightRect } from './HighlightRect';
-import { HighlightSelectionState, HighlightState, NO_SELECTION_STATE, SelectedState } from './HighlightState';
 import { getCssProperties } from './transformArea';
+import { HighlightState, HighlightStateType, NO_SELECTION_STATE } from './types/HighlightState';
 import type { RenderHighlightContentProps } from './types/RenderHighlightContentProps';
 import type { RenderHighlightsProps } from './types/RenderHighlightsProps';
 import type { RenderHighlightTargetProps } from './types/RenderHighlightTargetProps';
@@ -45,40 +45,39 @@ export const HighlightAreaList: React.FC<{
 
     // Filter the selections
     const listAreas =
-        highlightState instanceof HighlightSelectionState
+        highlightState.type === HighlightStateType.Selection
             ? highlightState.highlightAreas.filter((s) => s.pageIndex === pageIndex)
             : [];
 
     return (
         <>
             {renderHighlightTarget &&
-                highlightState instanceof SelectedState &&
+                (highlightState.type === HighlightStateType.Selected ||
+                    highlightState.type === HighlightStateType.ClickDragged) &&
                 highlightState.selectionRegion.pageIndex === pageIndex &&
                 renderHighlightTarget({
                     highlightAreas: highlightState.highlightAreas,
-                    selectedText: highlightState.selectedText,
+                    previewImage: highlightState.previewImage || '',
+                    selectedText: highlightState.selectedText || '',
                     selectionRegion: highlightState.selectionRegion,
                     selectionData: highlightState.selectionData,
                     cancel,
                     toggle: () => {
-                        store.update(
-                            'highlightState',
-                            new HighlightSelectionState(
-                                highlightState.selectedText,
-                                highlightState.highlightAreas,
-                                highlightState.selectionData,
-                                highlightState.selectionRegion
-                            )
-                        );
+                        // Switch to the `Selection` state
+                        const newState = Object.assign({}, highlightState, {
+                            type: HighlightStateType.Selection,
+                        });
+                        store.update('highlightState', newState);
                         window.getSelection().removeAllRanges();
                     },
                 })}
             {renderHighlightContent &&
-                highlightState instanceof HighlightSelectionState &&
+                highlightState.type == HighlightStateType.Selection &&
                 highlightState.selectionRegion.pageIndex === pageIndex &&
                 renderHighlightContent({
                     highlightAreas: highlightState.highlightAreas,
-                    selectedText: highlightState.selectedText,
+                    previewImage: highlightState.previewImage || '',
+                    selectedText: highlightState.selectedText || '',
                     selectionRegion: highlightState.selectionRegion,
                     selectionData: highlightState.selectionData,
                     cancel,