+++ packages/highlight/src/types/RenderHighlightTargetProps.ts
@@ -11,9 +11,10 @@ import type { SelectionData } from './SelectionData';
 
 export interface RenderHighlightTargetProps {
     highlightAreas: HighlightArea[];
+    previewImage: string;
     selectedText: string;
     selectionRegion: HighlightArea;
-    selectionData: SelectionData;
+    selectionData?: SelectionData;
     cancel(): void;
     // Switch to the hightlighting state
     toggle(): void;